--- android-json-form-wizard/src/main/java/com/vijay/jsonwizard/interfaces/GenericPopupInterface.java
@@ -1,11 +1,0 @@
-package com.vijay.jsonwizard.interfaces;
-
-import com.vijay.jsonwizard.utils.SecondaryValueModel;
-
-import java.util.Map;
-
-public interface GenericPopupInterface {
-    void onGenericDataPass(Map<String, SecondaryValueModel> selectedValues, String parentKey, String stepName, String childKey);
-
-    void clearSecondaryFields();
-}
